backend/requirements.txt

fastapi
uvicorn[standard]
python-dotenv
pydantic
SQLAlchemy
alembic
psycopg[binary]
faiss-cpu
numpy
scikit-learn
rank-bm25
pdfminer.six
python-docx
jinja2
beautifulsoup4
lxml
readability-lxml
requests
loguru
